requirements.txt
streamlit
python-dotenv
openai
langchain
langchain-openai
pydantic
typing-extensions
scipy
numpy
plotly
